requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
plotly==5.22.0
pandas==2.2.2
google-cloud-bigquery==3.25.0
db-dtypes==1.3.0
python-dotenv==1.0.1
pyarrow==16.1.0
